refactor: eliminate org concept and consolidate into scope layer
- Replace orgContract with scopeMemberContract under /api/scope/* paths
- Rename createUserScope → createScope with admin bypass fix for reserved slugs
- Absorb org-service functions into scope-member-service
- Migrate all /api/org/* routes to /api/scope/{members,invite,leave}
- Update CLI to use new scope-centric API endpoints
- Remove org.ts contract, org-service.ts, and entire /api/org/ directory
- Update all CLI and web tests to use new endpoint paths

Closes #3872
