+++ turbo/apps/cli/src/commands/scope/members.ts
@@ -1,15 +1,15 @@
 import { Command } from "commander";
 import chalk from "chalk";
-import { getOrgStatus } from "../../lib/api";
+import { getScopeMembers } from "../../lib/api";
 
 export const membersCommand = new Command()
   .name("members")
   .description("View scope members")
   .action(async () => {
     try {
-      const status = await getOrgStatus();
+      const status = await getScopeMembers();
 
-      console.log(chalk.bold(`Organization: ${status.slug}`));
+      console.log(chalk.bold(`Scope: ${status.slug}`));
       console.log(`  Role: ${status.role}`);
       console.log(
         `  Created: ${new Date(status.createdAt).toLocaleDateString()}`,
@@ -25,7 +25,7 @@ export const membersCommand = new Command()
       }
     } catch (error) {
       if (error instanceof Error) {
-        if (error.message.includes("Organization access token required")) {
+        if (error.message.includes("Scope access token required")) {
           console.error(
             chalk.red("✗ No active scope selected. Run: vm0 scope use <slug>"),
           );
