--- turbo/apps/web/app/api/org/__tests__/leave.test.ts
@@ -1,84 +1,0 @@
-import { describe, it, expect, beforeEach } from "vitest";
-import { POST as createOrgRoute } from "../route";
-import { POST } from "../leave/route";
-import { createTestRequest } from "../../../../src/__tests__/api-test-helpers";
-import { testContext, uniqueId } from "../../../../src/__tests__/test-helpers";
-import { mockClerk } from "../../../../src/__tests__/clerk-mock";
-import { setupClerkOrgMock } from "../../../../src/__tests__/org-test-helpers";
-
-const context = testContext();
-
-describe("POST /api/org/leave - Leave Organization", () => {
-  beforeEach(() => {
-    context.setupMocks();
-  });
-
-  it("should require authentication", async () => {
-    mockClerk({ userId: null });
-
-    const request = createTestRequest(
-      "http://localhost:3000/api/org/leave?scope=test",
-      {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({}),
-      },
-    );
-    const response = await POST(request);
-    const data = await response.json();
-
-    expect(response.status).toBe(401);
-    expect(data.error.message).toContain("Not authenticated");
-  });
-
-  it("should require scope query parameter", async () => {
-    const userId = uniqueId("leave-user");
-    mockClerk({ userId });
-
-    const request = createTestRequest("http://localhost:3000/api/org/leave", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({}),
-    });
-    const response = await POST(request);
-    const data = await response.json();
-
-    expect(response.status).toBe(400);
-    expect(data.error.message).toContain("scope query parameter is required");
-  });
-
-  it("should prevent admin from leaving", async () => {
-    const userId = uniqueId("leave-admin");
-    const slug = uniqueId("org");
-    const orgId = `org_${userId}`;
-    setupClerkOrgMock({
-      userId,
-      orgId,
-      memberships: [{ userId, role: "org:admin" }],
-    });
-
-    // Create org (fresh user, no existing scope)
-    const createReq = createTestRequest("http://localhost:3000/api/org", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ slug }),
-    });
-    const createRes = await createOrgRoute(createReq);
-    expect(createRes.status).toBe(201);
-
-    // Try to leave as admin
-    const leaveReq = createTestRequest(
-      `http://localhost:3000/api/org/leave?scope=${slug}`,
-      {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({}),
-      },
-    );
-    const leaveRes = await POST(leaveReq);
-    expect(leaveRes.status).toBe(403);
-
-    const leaveData = await leaveRes.json();
-    expect(leaveData.error.message).toContain("Admin");
-  });
-});
