--- turbo/apps/web/app/api/org/status/route.ts
@@ -1,49 +1,0 @@
-import { NextResponse } from "next/server";
-import { initServices } from "../../../../src/lib/init-services";
-import { getUserId } from "../../../../src/lib/auth/get-user-id";
-import { requireScopeFromRequest } from "../../../../src/lib/scope/resolve-scope";
-import { getOrganizationStatus } from "../../../../src/lib/org/org-service";
-import {
-  isBadRequest,
-  isNotFound,
-  isForbidden,
-} from "../../../../src/lib/errors";
-
-export async function GET(request: Request) {
-  initServices();
-
-  const authHeader = request.headers.get("authorization") ?? undefined;
-  const userId = await getUserId(authHeader);
-  if (!userId) {
-    return NextResponse.json(
-      { error: { message: "Not authenticated", code: "UNAUTHORIZED" } },
-      { status: 401 },
-    );
-  }
-
-  try {
-    const { scope } = await requireScopeFromRequest(request, userId);
-    const status = await getOrganizationStatus(userId, scope.id);
-    return NextResponse.json(status);
-  } catch (error) {
-    if (isBadRequest(error)) {
-      return NextResponse.json(
-        { error: { message: error.message, code: "BAD_REQUEST" } },
-        { status: 400 },
-      );
-    }
-    if (isForbidden(error)) {
-      return NextResponse.json(
-        { error: { message: error.message, code: "FORBIDDEN" } },
-        { status: 403 },
-      );
-    }
-    if (isNotFound(error)) {
-      return NextResponse.json(
-        { error: { message: error.message, code: "NOT_FOUND" } },
-        { status: 404 },
-      );
-    }
-    throw error;
-  }
-}
