+++ turbo/apps/web/app/api/scope/members/route.ts
@@ -2,12 +2,54 @@ import { NextResponse } from "next/server";
 import { initServices } from "../../../../src/lib/init-services";
 import { getUserId } from "../../../../src/lib/auth/get-user-id";
 import { requireScopeFromRequest } from "../../../../src/lib/scope/resolve-scope";
-import { removeMember } from "../../../../src/lib/org/org-service";
 import {
+  getScopeMembers,
+  removeMember,
+} from "../../../../src/lib/scope/scope-member-service";
+import {
+  isBadRequest,
   isNotFound,
   isForbidden,
-  isBadRequest,
 } from "../../../../src/lib/errors";
+
+export async function GET(request: Request) {
+  initServices();
+
+  const authHeader = request.headers.get("authorization") ?? undefined;
+  const userId = await getUserId(authHeader);
+  if (!userId) {
+    return NextResponse.json(
+      { error: { message: "Not authenticated", code: "UNAUTHORIZED" } },
+      { status: 401 },
+    );
+  }
+
+  try {
+    const { scope } = await requireScopeFromRequest(request, userId);
+    const status = await getScopeMembers(userId, scope.id);
+    return NextResponse.json(status);
+  } catch (error) {
+    if (isBadRequest(error)) {
+      return NextResponse.json(
+        { error: { message: error.message, code: "BAD_REQUEST" } },
+        { status: 400 },
+      );
+    }
+    if (isForbidden(error)) {
+      return NextResponse.json(
+        { error: { message: error.message, code: "FORBIDDEN" } },
+        { status: 403 },
+      );
+    }
+    if (isNotFound(error)) {
+      return NextResponse.json(
+        { error: { message: error.message, code: "NOT_FOUND" } },
+        { status: 404 },
+      );
+    }
+    throw error;
+  }
+}
 
 export async function DELETE(request: Request) {
   initServices();
@@ -33,7 +75,7 @@ export async function DELETE(request: Request) {
     const { scope, member } = await requireScopeFromRequest(request, userId);
     await removeMember(userId, scope.id, member.role, body.email);
     return NextResponse.json({
-      message: `Removed ${body.email} from organization`,
+      message: `Removed ${body.email} from scope`,
     });
   } catch (error) {
     if (isBadRequest(error)) {
