+++ turbo/apps/web/src/lib/telegram/handlers/shared.ts
@@ -6,7 +6,7 @@ import { agentComposes } from "../../../db/schema/agent-compose";
 import { getPlatformUrl } from "../../url";
 import {
   getUserScopeByClerkId,
-  createUserScope,
+  createScope,
   generateDefaultScopeSlug,
 } from "../../scope/scope-service";
 import { validateAgentSession } from "../../run";
@@ -222,10 +222,7 @@ async function completePendingLink(
 export async function ensureScopeAndArtifact(vm0UserId: string): Promise<void> {
   let scope = await getUserScopeByClerkId(vm0UserId);
   if (!scope) {
-    scope = await createUserScope(
-      vm0UserId,
-      generateDefaultScopeSlug(vm0UserId),
-    );
+    scope = await createScope(vm0UserId, generateDefaultScopeSlug(vm0UserId));
     log.info("Auto-created scope for Telegram user", { userId: vm0UserId });
   }
